Cast to int for union tag
Union tag will be changed from `enum` to `enum class`. We need explicit
casting with enum class.

Bug: 218912230
Test: m

+++ sensors/aidl/default/multihal/ConvertUtils.cpp
@@ -173,7 +173,7 @@ void convertToHidlEvent(const AidlEvent& aidlEvent, V2_1Event* hidlEvent) {
                 }
                 default:
                     ALOGE("Invalid sensor additioanl info tag: %d",
-                          additionalInfo.payload.getTag());
+                          static_cast<int32_t>(additionalInfo.payload.getTag()));
                     break;
             }
             break;
